Refactor the step definitions.
