+++ features/step_definitions/support/fixtures/user-fixtures.js
@@ -2,10 +2,9 @@ import {log, fail, getBrowser, getServer} from '../utils';
 
 fixtures.users = {
 
-  /* note: this is synchronous */
   create(name) {
     return getServer().execute((name) => {
-      const slug = require('/lib/utils').convertToSlug(name).replace(/-+/, '');
+      const slug = require('/lib/utils').convertToSlug(name).replace(/-+/, ''); // argh
 
       const user = Meteor.users.findOne(Accounts.createUser({
         email: slug + '@democracy.earth',
@@ -21,61 +20,6 @@ fixtures.users = {
 
       return user;
     }, name);
-  },
-
-  /* failing ; using createUser from server is ... a puzzle. Promises can't be serialized apparently. */
-  createWithServer(name) {
-    return getServer().execute((name) => {
-      const slug = require('/lib/utils').convertToSlug(name).replace(/-+/, '');
-
-      const returned = require('/imports/startup/both/modules/User.js').createUser({
-        email: slug + '@democracy.earth',
-        username: slug,
-        password: name,
-        mismatchPassword: name,
-      });
-      if ( ! returned) { throw new Error("User creation with createUser failed."); }
-
-      return returned;
-    }, name);
-  },
-
-  /* failing ; using createUser from browser fails to write to the profile because of security measures */
-  createWithBrowser(name) {
-    let slug = getServer().execute((name) => {
-      return require('/lib/utils').convertToSlug(name).replace(/-+/, '');
-    }, name);
-
-    getBrowser().timeoutsAsyncScript(3000);
-    const thing = getBrowser().executeAsync((name, slug, done) => {
-
-      const returned = require('/imports/startup/both/modules/User.js').createUser({
-        email: slug + '@democracy.earth',
-        username: slug,
-        password: name,
-        mismatchPassword: name,
-      });
-      if ( ! returned) { throw new Error("User creation with createUser failed."); }
-
-      returned.then((u) => {done(u)});
-      // executeAsync cannot return an Error, instead we get {}, so we wrap it ourselves
-      returned.catch((e) => {done({error: e.message})});
-
-    }, name, slug);
-
-    if (thing.value.error) { throw Error(thing.value.error); }
-
-    // Interesting. executeAsync cannot return an Error, instead we get {}
-    //if (thing.value instanceof Error) throw thing.value;
-
-    const user = getServer().execute((slug) => {
-      return Meteor.users.findOne({username: slug}); // use findOneByUsername instead
-    }, slug);
-
-    if ( ! user.profile) { fail(`No profile for user '${slug}' right after creation.`); }
-    if ( ! user) { fail(`No user '${slug}' in the database right after creation.`); }
-
-    return user;
   },
 
   login(email, password) {
@@ -97,23 +41,6 @@ fixtures.users = {
     if (returned && returned.value) {
       fail(`There was an error with the Meteor login : ${returned.value.message}.`);
     }
-  },
-
-
-  // These two do not belong here in the fixtures. Where ?
-
-  findByName(name) {
-    const user = getServer().execute((name) => {
-      const slug = require('/lib/utils').convertToSlug(name).replace(/-+/, '');
-      return require('meteor/accounts-base').Accounts.findUserByUsername(slug);
-    }, name);
-    return user;
-  },
-
-  findOneByName(name) {
-    const user = fixtures.users.findByName(name);
-    if ( ! user) { fail(`Unable to find the user '${name}'.`); }
-    return user;
   },
 
 };
