+++ features/step_definitions/support/models/_models.js
@@ -1,0 +1,1 @@
+global.models = {};
